requirements.txt
Flask>=3.0,<4
psycopg[binary]==3.2.3
psycopg-pool==3.2.6
gunicorn==22,<24
python-dotenv==0.21.0
urllib3==2.6.3
requests>=2.31.0